+++ examples/v1alpha3/nas/enas-cnn-cifar10/Dockerfile.gpu
@@ -23,10 +23,10 @@ RUN apt-get update && apt-get install -y software-properties-common && \
 
 
 ADD . /usr/src/app/github.com/kubeflow/katib
-WORKDIR /usr/src/app/github.com/kubeflow/katib/examples/v1alpha3/NAS-training-containers/RL-cifar10
+WORKDIR /usr/src/app/github.com/kubeflow/katib/examples/v1alpha3/nas/enas-cnn-cifar10
 
 RUN pip3 install --upgrade pip
 RUN pip3 install --no-cache-dir -r requirements-gpu.txt
-ENV PYTHONPATH /usr/src/app/github.com/kubeflow/katib/examples/v1alpha3/NAS-training-containers/RL-cifar10
+ENV PYTHONPATH /usr/src/app/github.com/kubeflow/katib/examples/v1alpha3/nas/enas-cnn-cifar10
 
 ENTRYPOINT ["python3.5", "-u", "RunTrial.py"]
