+++ examples/v1alpha3/nas/enas-example-cpu.yaml
@@ -36,7 +36,7 @@ spec:
             spec:
               containers:
               - name: {{.Trial}}
-                image: docker.io/kubeflowkatib/enas-cifar10-cpu
+                image: docker.io/kubeflowkatib/enas-cnn-cifar10-cpu
                 command:
                 - "python3.5"
                 - "-u"
