+++ prow_config.yaml
@@ -29,7 +29,7 @@ workflows:
     - cmd/suggestion/chocolate/v1alpha3/*
     - cmd/suggestion/hyperband/v1alpha3/*
     - cmd/suggestion/hyperopt/v1alpha3/*
-    - cmd/suggestion/nasrl/v1alpha3/*
+    - cmd/suggestion/nas/enas/v1alpha3/*
     - cmd/suggestion/skopt/v1alpha3/*
     - cmd/ui/v1alpha3/*
     - examples/v1alpha3/*.yaml
@@ -69,7 +69,7 @@ workflows:
     - cmd/suggestion/chocolate/v1alpha3/*
     - cmd/suggestion/hyperband/v1alpha3/*
     - cmd/suggestion/hyperopt/v1alpha3/*
-    - cmd/suggestion/nasrl/v1alpha3/*
+    - cmd/suggestion/nas/enas/v1alpha3/*
     - cmd/suggestion/skopt/v1alpha3/*
     - cmd/ui/v1alpha3/*
     - examples/v1alpha3/*.yaml
